Initial spigot commit.
