+++ shared/src/main/doc/LICENSE
