+++ shared/src/main/doc/README
