+++ shared/src/main/java/com/splunk/logtosplunk/Constants.java
@@ -1,0 +1,9 @@
+package com.splunk.logtosplunk;
+
+
+public class Constants {
+    public static final int MAX_PLAYERS = 128;
+
+    private Constants(){
+    }
+}
