+++ shared/src/main/java/com/splunk/logtosplunk/Point3dLong.java
