+++ shared/src/main/java/com/splunk/logtosplunk/SplunkConnection.java
