+++ shared/src/main/java/com/splunk/logtosplunk/SplunkMessagePreparer.java
