+++ shared/src/main/java/com/splunk/logtosplunk/loggable_events/LoggableEventType.java
