+++ shared/src/main/resources/mcmod.info
