+++ shared/src/test/java/com/splunk/logtosplunk/SplunkConnectionSpy.java
