+++ shared/src/test/java/com/splunk/logtosplunk/SplunkMessagePreparerSpy.java
